pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "research-trend-bot"
version = "0.1.0"
description = "Automated research paper digest bot using arxiv + Gemini"
requires-python = ">=3.9"
dependencies = [
    "google-genai>=1.0",
    "arxiv>=2.1",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "httpx>=0.27",
    "tenacity>=8.0",
    "pypdf>=4.0",
    "python-dotenv>=1.0",
]

[project.scripts]
research-trend-bot = "research_trend_bot.main:main"

[tool.setuptools.packages.find]
where = ["src"]
